requirements.txt
asgiref==3.3.1
awesomeversion==21.2.3
beautifulsoup4
cryptography==3.3.2
cloudinary
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
django-rest-auth[with_social]
djangorestframework
django-filter
django-datatables-view
gunicorn==20.1.0
markdown
Pillow
pyimgur
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.9.1
whitenoise


